refactor: remove makeLoader re-export
