+++ app/root.tsx
@@ -26,11 +26,8 @@ import { trpc } from "./trpc/client";
 import { publicConfig } from "./utils/config";
 import { ConfigPlaceholder } from "./utils/config-placeholder";
 import { useFlashMessages } from "./utils/flash-message-hook";
-import {
-  RootLoaderData,
-  makeLoader,
-  useRootLoaderData,
-} from "./utils/loader-utils";
+import { RootLoaderData, useRootLoaderData } from "./utils/loader-utils";
+import { makeLoader } from "./utils/loader-utils.server";
 import { cls } from "./utils/misc";
 import type { PageHandle } from "./utils/page-handle";
 import { QueryClientWrapper } from "./utils/react-query-utils";
@@ -48,7 +45,7 @@ export const links: LinksFunction = () => {
 // loader
 //
 
-export const loader = makeLoader(async ({ ctx }) => {
+export const loader = /* @__PURE__ */ makeLoader(async ({ ctx }) => {
   const loaderData: RootLoaderData = {
     currentUser: await ctx.currentUser(),
     flashMessages: await ctx.getFlashMessages(),
