+++ app/routes/decks/$id/export.tsx
@@ -1,7 +1,7 @@
 import { json } from "@remix-run/server-runtime";
 import { exportDeckJson } from "../../../misc/seed-utils";
 import { requireUserAndDeckV2 } from "../../../utils/loader-deck-utils";
-import { makeLoader } from "../../../utils/loader-utils";
+import { makeLoader } from "../../../utils/loader-utils.server";
 
 export const loader = /* @__PURE__ */ makeLoader(async ({ ctx }) => {
   const { deck } = await requireUserAndDeckV2(ctx);
