+++ app/routes/index.tsx
@@ -1,4 +1,5 @@
-import { makeLoader, useLoaderDataExtra } from "../utils/loader-utils";
+import { useLoaderDataExtra } from "../utils/loader-utils";
+import { makeLoader } from "../utils/loader-utils.server";
 import type { PageHandle } from "../utils/page-handle";
 import { PAGINATION_PARAMS_SCHEMA } from "../utils/pagination";
 import {
