+++ app/routes/videos/$id.tsx
@@ -24,11 +24,11 @@ import { trpc } from "../../trpc/client";
 import { useDocumentEvent } from "../../utils/hooks-client-utils";
 import { intl } from "../../utils/intl";
 import {
-  makeLoader,
   useLeafLoaderData,
   useLoaderDataExtra,
   useRootLoaderData,
 } from "../../utils/loader-utils";
+import { makeLoader } from "../../utils/loader-utils.server";
 import { cls } from "../../utils/misc";
 import type { PageHandle } from "../../utils/page-handle";
 import type { CaptionEntry } from "../../utils/types";
